+++ src/test/java/com/iexec/worker/executor/TaskManagerServiceTests.java
@@ -849,7 +849,7 @@ class TaskManagerServiceTests {
         ReplicateActionResponse replicateActionResponse =
                 taskManagerService.contribute(CHAIN_TASK_ID);
 
-        Assertions.assertThat(replicateActionResponse)
+        assertThat(replicateActionResponse)
                 .isNotNull()
                 .isEqualTo(ReplicateActionResponse
                         .success(chainReceipt));
@@ -863,10 +863,9 @@ class TaskManagerServiceTests {
         ReplicateActionResponse replicateActionResponse =
                 taskManagerService.contribute(CHAIN_TASK_ID);
 
-        Assertions.assertThat(replicateActionResponse)
+        assertThat(replicateActionResponse)
                 .isNotNull()
-                .isEqualTo(ReplicateActionResponse
-                        .failure(CONTRIBUTION_TIMEOUT));
+                .isEqualTo(ReplicateActionResponse.failure(CONTRIBUTION_TIMEOUT));
     }
 
     @Test
@@ -1178,6 +1177,112 @@ class TaskManagerServiceTests {
     }
     //endregion
 
+    // region contributeAndFinalize
+    @Test
+    void shouldNotContributeAndFinalizeSinceCannotContributeStatusIsPresent() {
+        when(contributionService.getCannotContributeStatusCause(CHAIN_TASK_ID))
+                .thenReturn(Optional.of(CONTRIBUTION_TIMEOUT));
+        ReplicateActionResponse replicateActionResponse = taskManagerService.contributeAndFinalize(CHAIN_TASK_ID);
+        assertThat(replicateActionResponse)
+                .isNotNull()
+                .isEqualTo(ReplicateActionResponse.failure(CONTRIBUTION_TIMEOUT));
+    }
+
+    @Test
+    void shouldNotContributeAndFinalizeSinceNotEnoughGas() {
+        when(contributionService.getCannotContributeStatusCause(CHAIN_TASK_ID)).thenReturn(Optional.empty());
+        when(iexecHubService.hasEnoughGas()).thenReturn(false);
+        ReplicateActionResponse replicateActionResponse = taskManagerService.contributeAndFinalize(CHAIN_TASK_ID);
+        assertThat(replicateActionResponse)
+                .isNotNull()
+                .isEqualTo(ReplicateActionResponse.failure(OUT_OF_GAS));
+    }
+
+    @Test
+    void shouldNotContributeAndFinalizeSinceNoComputedFile() {
+        when(contributionService.getCannotContributeStatusCause(CHAIN_TASK_ID)).thenReturn(Optional.empty());
+        when(iexecHubService.hasEnoughGas()).thenReturn(true);
+        when(resultService.getComputedFile(CHAIN_TASK_ID)).thenReturn(null);
+        ReplicateActionResponse replicateActionResponse = taskManagerService.contributeAndFinalize(CHAIN_TASK_ID);
+        assertThat(replicateActionResponse)
+                .isNotNull()
+                .isEqualTo(ReplicateActionResponse.failure(DETERMINISM_HASH_NOT_FOUND));
+    }
+
+    @Test
+    void shouldNotContributeAndFinalizeSinceNoContribution() {
+        ComputedFile computedFile = ComputedFile.builder()
+                .resultDigest("digest")
+                .callbackData("data")
+                .build();
+        when(contributionService.getCannotContributeStatusCause(CHAIN_TASK_ID)).thenReturn(Optional.empty());
+        when(iexecHubService.hasEnoughGas()).thenReturn(true);
+        when(resultService.getComputedFile(CHAIN_TASK_ID)).thenReturn(computedFile);
+        when(contributionService.getContribution(computedFile)).thenReturn(null);
+        ReplicateActionResponse replicateActionResponse = taskManagerService.contributeAndFinalize(CHAIN_TASK_ID);
+        assertThat(replicateActionResponse)
+                .isNotNull()
+                .isEqualTo(ReplicateActionResponse.failure(ENCLAVE_SIGNATURE_NOT_FOUND));
+    }
+
+    @Test
+    void shouldNotContributeAndFinalizeSinceTransactionFailed() {
+        ComputedFile computedFile = ComputedFile.builder()
+                .resultDigest("digest")
+                .callbackData("data")
+                .build();
+        Contribution contribution = mock(Contribution.class);
+        when(contributionService.getCannotContributeStatusCause(CHAIN_TASK_ID)).thenReturn(Optional.empty());
+        when(iexecHubService.hasEnoughGas()).thenReturn(true);
+        when(resultService.getComputedFile(CHAIN_TASK_ID)).thenReturn(computedFile);
+        when(contributionService.getContribution(computedFile)).thenReturn(contribution);
+        when(iexecHubService.contributeAndFinalize(any(), any(), any())).thenReturn(Optional.empty());
+        ReplicateActionResponse replicateActionResponse = taskManagerService.contributeAndFinalize(CHAIN_TASK_ID);
+        assertThat(replicateActionResponse)
+                .isNotNull()
+                .isEqualTo(ReplicateActionResponse.failure(CHAIN_RECEIPT_NOT_VALID));
+    }
+
+    @Test
+    void shouldNotContributeAndFinalizeSinceInvalidChainReceipt() {
+        ComputedFile computedFile = ComputedFile.builder()
+                .resultDigest("digest")
+                .callbackData("data")
+                .build();
+        ChainReceipt chainReceipt = ChainReceipt.builder().blockNumber(0).build();
+        Contribution contribution = mock(Contribution.class);
+        when(contributionService.getCannotContributeStatusCause(CHAIN_TASK_ID)).thenReturn(Optional.empty());
+        when(iexecHubService.hasEnoughGas()).thenReturn(true);
+        when(resultService.getComputedFile(CHAIN_TASK_ID)).thenReturn(computedFile);
+        when(contributionService.getContribution(computedFile)).thenReturn(contribution);
+        when(iexecHubService.contributeAndFinalize(any(), any(), any())).thenReturn(Optional.of(chainReceipt));
+        ReplicateActionResponse replicateActionResponse = taskManagerService.contributeAndFinalize(CHAIN_TASK_ID);
+        assertThat(replicateActionResponse)
+                .isNotNull()
+                .isEqualTo(ReplicateActionResponse.failure(CHAIN_RECEIPT_NOT_VALID));
+    }
+
+    @Test
+    void shouldContributeAndFinalize() {
+        ComputedFile computedFile = ComputedFile.builder()
+                .resultDigest("digest")
+                .callbackData("data")
+                .build();
+        Contribution contribution = mock(Contribution.class);
+        ChainReceipt chainReceipt = ChainReceipt.builder().blockNumber(10).build();
+        when(contributionService.getCannotContributeStatusCause(CHAIN_TASK_ID)).thenReturn(Optional.empty());
+        when(iexecHubService.hasEnoughGas()).thenReturn(true);
+        when(resultService.getComputedFile(CHAIN_TASK_ID)).thenReturn(computedFile);
+        when(contributionService.getContribution(computedFile)).thenReturn(contribution);
+        when(resultService.uploadResultAndGetLink(CHAIN_TASK_ID)).thenReturn("resultLink");
+        when(iexecHubService.contributeAndFinalize(any(), anyString(), anyString())).thenReturn(Optional.of(chainReceipt));
+        ReplicateActionResponse replicateActionResponse = taskManagerService.contributeAndFinalize(CHAIN_TASK_ID);
+        assertThat(replicateActionResponse)
+                .isNotNull()
+                .isEqualTo(ReplicateActionResponse.success("resultLink", "data"));
+    }
+    // endregion
+
     //region complete
     @Test
     void shouldComplete() {
@@ -1240,8 +1345,10 @@ class TaskManagerServiceTests {
 
     @Test
     void shouldFindChainReceiptValid() {
-        Optional<ChainReceipt> receipt = Optional.of(new ChainReceipt(5,
-                "txHash"));
+        ChainReceipt receipt = ChainReceipt.builder()
+                .blockNumber(5)
+                .txHash("txHash")
+                .build();
 
         boolean isValid =
                 taskManagerService.isValidChainReceipt(CHAIN_TASK_ID, receipt);
@@ -1249,9 +1356,16 @@ class TaskManagerServiceTests {
     }
 
     @Test
+    void shouldFindChainReceiptNotValidSinceNull() {
+        assertThat(taskManagerService.isValidChainReceipt(CHAIN_TASK_ID, null)).isFalse();
+    }
+
+    @Test
     void shouldFindChainReceiptNotValidSinceBlockIsZero() {
-        Optional<ChainReceipt> receipt = Optional.of(new ChainReceipt(0,
-                "txHash"));
+        ChainReceipt receipt = ChainReceipt.builder()
+                .blockNumber(0)
+                .txHash("txHash")
+                .build();
 
         boolean isValid =
                 taskManagerService.isValidChainReceipt(CHAIN_TASK_ID, receipt);