+++ src/test/java/com/iexec/worker/executor/TaskNotificationServiceTest.java
@@ -318,6 +318,48 @@ class TaskNotificationServiceTest {
     }
 
     @Test
+    void shouldContributeAndFinalize() throws InterruptedException {
+        TaskNotification currentNotification = TaskNotification.builder()
+                .chainTaskId(CHAIN_TASK_ID)
+                .taskNotificationType(PLEASE_CONTRIBUTE_AND_FINALIZE)
+                .build();
+        when(iexecHubService.getTaskDescription(CHAIN_TASK_ID)).thenReturn(taskDescription);
+        when(taskManagerService.contributeAndFinalize(CHAIN_TASK_ID)).thenReturn(ReplicateActionResponse.success());
+        when(iexecHubService.getChainTask(CHAIN_TASK_ID)).thenReturn(Optional.of(getChainTask()));
+        doNothing().when(subscriptionService).waitForSessionReady();
+        when(customCoreFeignClient.updateReplicateStatus(anyString(), any())).thenReturn(PLEASE_WAIT);
+
+        taskNotificationService.onTaskNotification(currentNotification);
+        verify(taskManagerService).contributeAndFinalize(CHAIN_TASK_ID);
+        TaskNotification nextNotification = TaskNotification.builder()
+                .chainTaskId(CHAIN_TASK_ID)
+                .taskNotificationType(PLEASE_WAIT)
+                .build();
+        verify(applicationEventPublisher).publishEvent(nextNotification);
+    }
+
+    @Test
+    void shouldNotContributeAndFinalize() throws InterruptedException {
+        TaskNotification currentNotification = TaskNotification.builder()
+                .chainTaskId(CHAIN_TASK_ID)
+                .taskNotificationType(PLEASE_CONTRIBUTE_AND_FINALIZE)
+                .build();
+        when(iexecHubService.getTaskDescription(CHAIN_TASK_ID)).thenReturn(taskDescription);
+        when(taskManagerService.contributeAndFinalize(CHAIN_TASK_ID)).thenReturn(ReplicateActionResponse.failure());
+        when(iexecHubService.getChainTask(CHAIN_TASK_ID)).thenReturn(Optional.of(getChainTask()));
+        doNothing().when(subscriptionService).waitForSessionReady();
+        when(customCoreFeignClient.updateReplicateStatus(anyString(), any())).thenReturn(PLEASE_ABORT);
+
+        taskNotificationService.onTaskNotification(currentNotification);
+        verify(taskManagerService).contributeAndFinalize(CHAIN_TASK_ID);
+        TaskNotification nextNotification = TaskNotification.builder()
+                .chainTaskId(CHAIN_TASK_ID)
+                .taskNotificationType(PLEASE_ABORT)
+                .build();
+        verify(applicationEventPublisher).publishEvent(nextNotification);
+    }
+
+    @Test
     void shouldReveal() throws InterruptedException {
         when(iexecHubService.getTaskDescription(CHAIN_TASK_ID)).thenReturn(taskDescription);
         TaskNotification currentNotification = TaskNotification.builder().chainTaskId(CHAIN_TASK_ID)
@@ -370,7 +412,7 @@ class TaskNotificationServiceTest {
     }
 
     @Test
-    void shouldComplete() throws InterruptedException {
+    void shouldComplete() {
         when(iexecHubService.getTaskDescription(CHAIN_TASK_ID)).thenReturn(taskDescription);
         TaskNotification currentNotification = TaskNotification.builder().chainTaskId(CHAIN_TASK_ID)
                 .taskNotificationType(PLEASE_COMPLETE)