+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### New Features
 - Enable Prometheus actuator. (#512)
-- Add `contributeAndFinalize` support for TEE tasks when required `trust` is `1`. (#516)
+- Add `contributeAndFinalize` support for TEE tasks when required `trust` is `1`. (#516 #517)
 ### Bug Fixes
 - Filter `contribute` and `reveal` event logs  on `chainTaskId` and `workerAddress`. (#516)
 ### Quality